login refactored
