+++ pages/side-panel/src/ChatInterface.tsx
@@ -42,6 +42,17 @@ export interface Message {
   };
 }
 
+// User type definition
+type User = {
+  id: string;
+  email?: string;
+  user_metadata?: {
+    avatar_url?: string;
+    full_name?: string;
+    [key: string]: unknown;
+  };
+} | null;
+
 interface ChatInterfaceProps {
   selectedModel: string;
   setSelectedModel: (model: string) => void;
@@ -54,6 +65,10 @@ interface ChatInterfaceProps {
   anthropicModels: Array<{ name: string; displayName?: string; provider: string }>;
   isLoadingModels?: boolean;
   modelError?: string | null;
+  // User state and authentication handlers
+  user: User;
+  handleSignIn: () => Promise<void>;
+  handleSignOut: () => Promise<void>;
 }
 
 export const ChatInterface = forwardRef<
@@ -76,6 +91,9 @@ export const ChatInterface = forwardRef<
     anthropicModels,
     isLoadingModels,
     modelError,
+    user,
+    handleSignIn,
+    handleSignOut,
   } = props;
 
   const fontFamily = useStorage(fontFamilyStorage);
@@ -483,7 +501,7 @@ export const ChatInterface = forwardRef<
       onDragOver={handleDragOver}
       onDragLeave={handleDragLeave}
       onDrop={handleDrop}>
-      <HeaderComponent />
+      <HeaderComponent user={user} handleSignIn={handleSignIn} handleSignOut={handleSignOut} />
 
       {/* Use context messages directly as our source of truth */}
       <MessageList
